+++ app/scripts/controllers/localmap.js
@@ -157,10 +157,8 @@ angular.module('accessimapEditeurDerApp')
       $scope.styleChoices = settings.STYLES[$scope.queryChosen.type];
       $scope.styleChosen = $scope.styleChoices[0];
 
-      $scope.colors = settings.COLORS.other;
+      $scope.colors = (settings.COLORS.transparent).concat(settings.COLORS.other);
       $scope.colorChosen = $scope.colors[0];
-      console.log($scope.colorChosen)
-      console.log($scope.colors)
 
       $scope.changeStyle = function() {
           $scope.styleChoices = settings.STYLES[$scope.queryChosen.type];
@@ -276,11 +274,8 @@ angular.module('accessimapEditeurDerApp')
             var k = attribute.k;
             var v = attribute.v;
             if (k === 'fill-pattern') {
-              if ($scope.checkboxModel.fill) {
-                console.log($scope.colorChosen)
-                if ($scope.colorChosen) {
-                  v += '_' + $scope.colorChosen.color;
-                }
+              if ($scope.colorChosen && $scope.colorChosen.color !== 'none') {
+                v += '_' + $scope.colorChosen.color;
               }
               symbol.attr('fill', settings.POLYGON_STYLES[v].url());
             } else {
@@ -364,8 +359,8 @@ angular.module('accessimapEditeurDerApp')
           var k = attribute.k;
           var v = attribute.v;
           if (k === 'fill-pattern') {
-            if ($scope.geojson[objectId].fill) {
-              v += '_bg';
+            if ($scope.geojson[objectId].color && $scope.geojson[objectId].color.color !== 'none') {
+              v += '_' + $scope.geojson[objectId].color.color;
             }
             d3.select('#' + id)
               .attr('fill', settings.POLYGON_STYLES[v].url());
@@ -405,8 +400,8 @@ angular.module('accessimapEditeurDerApp')
           var k = attribute.k;
           var v = attribute.v;
           if (k === 'fill-pattern') {
-            if ($scope.geojson[objectId].fill) {
-              v += '_bg';
+            if ($scope.geojson[objectId].color && $scope.geojson[objectId].color.color !== 'none') {
+              v += '_' + $scope.geojson[objectId].color.color;
             }
             symbol.attr('fill', settings.POLYGON_STYLES[v].url());
           } else {
@@ -444,7 +439,6 @@ angular.module('accessimapEditeurDerApp')
       }
 
       function drawFeature(data, feature, optionalClass) {
-        console.log(data, feature, optionalClass)
         var featureGroup;
         if (optionalClass) {
           if (d3.select('.vector.' + optionalClass + '#' + feature[0].id).empty()) {
@@ -510,10 +504,8 @@ angular.module('accessimapEditeurDerApp')
           var k = attribute.k;
           var v = attribute.v;
           if (k === 'fill-pattern') {
-            if ($scope.checkboxModel.fill) {
-              if ($scope.colorChosen) {
-                v += '_' + $scope.colorChosen.color;
-              }
+            if ($scope.colorChosen && $scope.colorChosen.color !== 'none') {
+              v += '_' + $scope.colorChosen.color;
             }
             d3.select('#' + feature[0].id).attr('fill', settings.POLYGON_STYLES[v].url());
           } else {
@@ -525,8 +517,8 @@ angular.module('accessimapEditeurDerApp')
             var k = attribute.k;
             var v = attribute.v;
             if (k === 'fill-pattern') {
-              if ($scope.checkboxModel.fill) {
-                v += '_bg';
+              if ($scope.colorChosen && $scope.colorChosen.color !== 'none') {
+                v += '_' + $scope.colorChosen.color;
               }
               d3.select('.' + optionalClass + '#' + feature[0].id).attr('fill', settings.POLYGON_STYLES[v].url());
             } else {
@@ -592,7 +584,7 @@ angular.module('accessimapEditeurDerApp')
                 style: $scope.styleChosen,
                 styleChoices: $scope.styleChoices,
                 contour: $scope.checkboxModel.contour,
-                fill: $scope.checkboxModel.fill
+                color: $scope.colorChosen
               };
               addToLegend($scope.queryChosen, $scope.styleChosen, $scope.geojson.length);
             }